REF: reorganize flexible BSON code
