+++ bson/__init__.py
@@ -778,7 +778,7 @@ def _name_value_to_bson(name, value, check_keys, opts):
     # Finally, try to encode the object with registered codec.
     codec = opts.custom_codec_map.get(type(value))
     if codec is not None:
-        from bson.bson_writer import BSONFieldWriter
+        from bson.flexible_bson.writer import BSONFieldWriter
         writer = BSONFieldWriter(name)
         writer.start_document()
         codec.to_bson(value, writer)
