+++ bson/flexible_bson/__init__.py
@@ -1,0 +1,3 @@
+from bson.flexible_bson.writer import (
+    BSONWriter, BSONDocumentWriter, BSONFieldWriter)
+from bson.flexible_bson.codec import BSONCodecBase
