+++ bson/flexible_bson/writer.py
